work around Docker’s non-deterministic network interface assignment
The proxy is set up with two Docker networks. This creates two network
interfaces, eth0 and eth1, in the proxy container. However, the setup is
not deterministic: We don't know which interface will be assigned to
which network.

To restore determinisim, we therefore switch the names of eth0 and eth1
around such the eth0 is the interface facing the client, and eth1 the
interface facing the server.