work around Docker’s non-deterministic network interface assignment (#43)
The proxy is set up with two Docker networks. This creates two network interfaces, eth0 and eth1, in the proxy container. However, the setup is not deterministic: We don't know which interface will be assigned to which network.

To restore determinism, we therefore need to determine if eth0 is facing the client and eth1 is facing the server, or vice versa.